requirements.txt
Flask==3.1.2
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.11
Werkzeug==3.1.3
SQLAlchemy==2.0.44
python-dotenv==1.2.1
